Done with ExpensesChart

+++ budget-tracker/src/components/Expenses/ExpenseDate.js
@@ -1,7 +1,7 @@
 import './ExpenseDate.css'
 
 const ExpenseDate = (props) => {
-    const month = props.date.toLocaleString("en-us",{month:"long"});
+    const month = props.date.toLocaleString("en-us",{month:"short"});
     const day = props.date.toLocaleString("en-us",{day:"2-digit"});
     const year= props.date.getFullYear();
   return (
